To support newest VS 2017 RC3 change yield to co_yield Warnings fixed

+++ AdventOfCode2016/Libraries/md5.cpp
@@ -111,7 +111,7 @@ MD5::MD5()
 MD5::MD5(const std::string &text)
 {
 	init();
-	update(text.c_str(), text.length());
+	update(text.c_str(), static_cast<unsigned int>(text.length()));
 	finalize();
 }
